refactor: UserDetailService를 인터페이스로 만들어서 auth 패키지 의존성 완전히 제거
