+++ src/main/java/auth/AuthConfiguration.java
@@ -1,33 +1,25 @@
 package auth;
 
 import lombok.RequiredArgsConstructor;
-import nextstep.member.MemberDao;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @RequiredArgsConstructor
 @Configuration
 public class AuthConfiguration {
-    private final MemberDao memberDao;
-
     @Bean
     public JwtTokenProvider jwtTokenProvider() {
         return new JwtTokenProvider();
     }
 
     @Bean
-    public UserDetailService userDetailService() {
-        return new UserDetailService(memberDao);
+    public AuthenticationProvider authenticationProvider(UserDetailService userDetailService) {
+        return new AuthenticationProvider(jwtTokenProvider(), userDetailService);
     }
 
     @Bean
-    public AuthenticationProvider authenticationProvider() {
-        return new AuthenticationProvider(jwtTokenProvider(), userDetailService());
-    }
-
-    @Bean
-    public LoginController loginController() {
-        return new LoginController(authenticationProvider());
+    public LoginController loginController(AuthenticationProvider authenticationProvider) {
+        return new LoginController(authenticationProvider);
     }
 
     @Bean
@@ -36,7 +28,7 @@ public class AuthConfiguration {
     }
 
     @Bean
-    public LoginMemberArgumentResolver loginMemberArgumentResolver() {
-        return new LoginMemberArgumentResolver(jwtTokenProvider(), userDetailService());
+    public LoginMemberArgumentResolver loginMemberArgumentResolver(UserDetailService userDetailService) {
+        return new LoginMemberArgumentResolver(jwtTokenProvider(), userDetailService);
     }
 }
